Codeaza Internship - Final Project/requirements.txt
Flask==2.3.2
itemadapter==0.8.0
mysql_connector_repackaged==0.3.1
pandas==2.0.2
Scrapy==2.9.0
